feat: create student repository interface
